// ExistingCode/src/data-access/domain/tables/collar/schema.ts
export const CollarTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	collarId: z.any().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	parentCollarId: z.any().nullable().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	loggingEventId: z.any(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: nvarchar
	 */
	project: z.string().max(100).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	prospect: z.string().max(30).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	tenement: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	target: z.string().max(100).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	subTarget: z.string().max(100).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	pit: z.string().max(30).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	phase: z.string().max(30).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	section: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	holeType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	holePurpose: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	holePurposeDetail: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	collarType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	preCollarId: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	redox: z.string().max(50).nullable().optional(),
	/**
	 * dataType: decimal
	 */
	casingDepth: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	preCollarDepth: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	startDepth: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	totalDepth: z.number().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	startedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	finishedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	approvedInd: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	responsiblePerson: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	responsiblePerson2: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	explorationCompany: z.string().max(50).nullable().optional(),
	/**
	 * dataType: smallint
	 * defaultValue: ((0))
	 */
	priority: z.number().int().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	waterTableDepth: z.number().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	waterTableDepthMeasuredOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	orientationTool: z.string().max(50).nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	modelUseInd: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	dataSource: z.string().max(255),
	/**
	 * dataType: binary
	 */
	migrationRv: z.any().nullable().optional(),
	/**
	 * dataType: int
	 */
	migrationId: z.number().int().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(4000).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});